% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_package_VC_libPaths}
\alias{add_package_VC_libPaths}
\title{Append all package locations to `.libPaths()`.}
\usage{
add_package_VC_libPaths(packNameVersion, lib.location,
  additional_lib_paths = c())
}
\arguments{
\item{packNameVersion}{A named character vector with package names and their version indication (e.g. `c(dplyr = '>= 0.05', ggplot = '')`).
The path that is appended to the `.libPaths` is constructed based on the name and version provided.}

\item{lib.location}{The R_VC_library location path (no default configured here).}
}
\description{
Adds the path of the package that is specified (and likely loaded before) to the `.libPaths`.
}

