% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{availablePackageVersions}
\alias{availablePackageVersions}
\title{Just checks the R_VC_library folder structure for all available versions installed for a specific package.
If no name is provided, an error is returned.}
\usage{
availablePackageVersions(packageName, lib.location = R_VC_library_location())
}
\arguments{
\item{packageName}{The name of the package for which all versions must be returned.}

\item{lib.location}{The folder containing the structure where this package his versions need to be checked.}
}
\description{
Just checks the R_VC_library folder structure for all available versions installed for a specific package.
If no name is provided, an error is returned.
}

