% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{checkIfBasePackage}
\alias{checkIfBasePackage}
\title{Check if a package belongs to the standard R (base) packages.}
\usage{
checkIfBasePackage(packageName)
}
\arguments{
\item{packageName}{The package name to check.}
}
\description{
Check if a package belongs to the standard R (base) packages.
}
