% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{clean_download_catch}
\alias{clean_download_catch}
\title{Clean package download catch}
\usage{
clean_download_catch()
}
\value{
No return value, is called for it's side-effect of removing the \code{file.path(tempdir(), 'downloaded_packages')} folder.
}
\description{
Clean the catch folder 'downloaded_packages' which lives in the temporary R session folder `tempdir()`.
}
