% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{clean_temp_library}
\alias{clean_temp_library}
\title{Clear the temp install folder.}
\usage{
clean_temp_library(temp_install.location = R_VC_temp_lib_location())
}
\arguments{
\item{install.location}{By default the default temporary directory path obtained with \code{R_VC_temp_lib_location()}.}
}
\description{
The temporary installation folder (indicated by \code{R_VC_temp_lib_location()}) is used to install the package before moving ('converting') it to the final location.
This function removes this temporary folder. Make sure that all installed packages that are desired to keep are converted.
You can run the \code{convert_to_VC_library()} once again to make sure this is the case.
}

