% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{convert_to_VC_library}
\alias{convert_to_VC_library}
\title{Move normally installed packages to R_VC_library structure.}
\usage{
convert_to_VC_library(normalLibrary = R_VC_temp_lib_location(),
  VC_library_location = R_VC_library_location())
}
\arguments{
\item{normalLibrary}{The temporary library where a package was temporarily installed (having a normal library structure).
By default, it checks the environment variable \code{R_VC_LIBRARY_LOCATION} and appends \code{/TEMP_install_location}.}

\item{VC_library_location}{The folder containing a structure where all packages in the temp folder must be moved to.
By default, it checks the environment variable \code{R_VC_LIBRARY_LOCATION} for this directory.}
}
\description{
Move normally installed packages to R_VC_library structure.
}
