% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{dependsOnMe}
\alias{dependsOnMe}
\title{Shows the dependencies of (all or) a certain function(s).}
\usage{
dependsOnMe(..., checkMyDeps = NULL, lib.location = R_VC_library_location())
}
\arguments{
\item{...}{All packages and their versions you would like to load e.g. \code{dependsOnMe(DBI = '0.5', assertthat, R6 = '0.6', quietly = TRUE)}.}

\item{checkMyDeps}{Supports providing a named character vector of packages and their versions instead of the direct input.
Use it like this when calling it via another function. It is the shape that is supported by all other functions in this package.
e.g. \code{c(DBI = '0.5', assertthat = '', R6 = '')}}

\item{lib.location}{The folder containing a structure where this function observe the dependencies from. By default, it checks the environment variable \code{R_VC_LIBRARY_LOCATION} for this directory.}
}
\description{
Can be called without using quotes like \code{dependsOnMe(dplyr)}. A second very usefull feature would be the \code{dependsOnMe(all)},
which will print a list of all packages available with their dependencies. \cr
\cr
A simple wrapper to do precisely that is "\code{installed.packages_VC}".
}

