% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{error_packageAlreadyLoaded}
\alias{error_packageAlreadyLoaded}
\title{Throw error because this package is already loaded and not compatible with the requested version.}
\usage{
error_packageAlreadyLoaded(
  requested_package_name,
  requested_version,
  already_loaded_version
)
}
\arguments{
\item{requested_package_name}{A single package name of the package that was desired to be loaded.}

\item{requested_version}{The version definition (like: ">= 3.2.1") of the package that was tried to be loaded.}

\item{already_loaded_version}{The version of the package that is already loaded, which causes this error to be fired.}
}
\description{
Throw error because this package is already loaded and not compatible with the requested version.
}
