% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{execute_with_packages}
\alias{execute_with_packages}
\title{Alias of \code{\link{lib.execute_using_packagelist}}}
\usage{
execute_with_packages(packages_to_load = NULL, func_handle, ...,
  execution_log_location = normPath(paste0("./execution_logs/execution_",
  gsub("\\\\s|-|:", "_", format(Sys.time())), ".txt")),
  lib_location = lib.location(), wait_for_response = TRUE,
  execute_quietly = FALSE)
}
\description{
Alias of \code{\link{lib.execute_using_packagelist}}
}
