% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{getOnlineDependencies}
\alias{getOnlineDependencies}
\title{Check if an package belongs to the standard R packages}
\usage{
getOnlineDependencies(packageName, cran_url = "https://cran.rstudio.com/")
}
\arguments{
\item{packageName}{The package name to check.}
}
\description{
Check if an package belongs to the standard R packages
}

