% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{install.packages_VC}
\alias{install.packages_VC}
\title{Install CRAN package into VC library.}
\usage{
install.packages_VC(installPackages = NULL,
  lib.location = R_VC_library_location(), install_temporarily = FALSE,
  overwrite_this_package = FALSE, execute_with_Rscript = TRUE,
  verbose = FALSE)
}
\arguments{
\item{installPackages}{Provide a vector of package names. A version cannot be supplied.}

\item{lib.location}{The folder where this package can be installed.
The package will first be installed in a temporary install folder \code{yourLib.loc/TEMP_install_location}
indicated by the \code{R_VC_temp_lib_location} function. If \code{install_temporarily} is set to FALSE (the default).
the installed package(s) is moved to the destination location automatically.}

\item{install_temporarily}{If FALSE, the installed package(s) is (are) moved to the final destination automatically.
Otherwise it is necessary to run \code{convert_to_VC_library()} manually after the installation into the temporary folder finished. To clean up the temporary folder, run \code{clean_temp_library()}.
When \code{overwrite_this_package} is TRUE, and \code{install_temporarily} is not provided, it will be set to FALSE automatically.
Installing a package temporarily gives you the oppertunity to test the package before adding it to the RVClibrary structure.
Loading packages including those in the temporary library (\code{R_VC_temp_lib_location()}) can be done using: \code{library_VC(..., also_load_from_temp_lib = TRUE)}.}

\item{execute_with_Rscript}{When TRUE (the default), it will try to install the package using a sepparate Rscipt R instance.
This simplifies installing and leaves changing loaded packages etc. to another R instance which we can kill :).
It will run a script provided with your packages. That script will load \code{RVClibrary} and call \code{install.packages_VC} directly.}
}
\description{
This function can be used to install a CRAN package in the (default or specified) R_VC_library.
For installing packages from source use \code{install.packages_VC_tarball}.
}

