% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{installDevtools}
\alias{installDevtools}
\title{Installs the incredible `devtools` package.}
\usage{
installDevtools(lib.location = R_VC_library_location(),
  force_install = FALSE)
}
\arguments{
\item{lib.location}{The library (can be an empty folder) to install Devtools in.}

\item{force_install}{FALSE by default, if the package `devTools` is already installed, it will return silently.}
}
\description{
Will install devtools and it's dependencies into the library provided by `R_VC_library_location()`.
If the
}

