% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{installed.packages_VC}
\alias{installed.packages_VC}
\title{Show the complete library content.}
\usage{
installed.packages_VC()
}
\description{
Use to print all available packages in the VC_library with all their versions including their dependencies.
Simply performs a call to \code{dependsOnMe(all)}.
}
