% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{isVersionCompatible}
\alias{isVersionCompatible}
\title{check if version indication is compliant.}
\usage{
isVersionCompatible(condition, version)
}
\arguments{
\item{condition}{A version indication like `>= 4.5.1` or `2.3.4` or `> 1.2.3` or `''` (empty) or `NA`.}

\item{version}{A version number like `1.2.3`.}
}
\description{
Returns TRUE if the 'condition' complies with the already available 'version'.
}

