% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{lib.available_versions}
\alias{lib.available_versions}
\title{Just checks the multiversion library for all available versions installed for a specific package.
If no name is provided, an error is returned.}
\usage{
lib.available_versions(packageName, lib_location = lib.location())
}
\arguments{
\item{packageName}{The name of the package for which all versions must be returned.}

\item{lib_location}{The folder containing the structure where this package his versions need to be checked.}
}
\value{
A character vector with the different versions that are available for a specific package.
}
\description{
Just checks the multiversion library for all available versions installed for a specific package.
If no name is provided, an error is returned.
}
