% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{lib.check_compatibility}
\alias{lib.check_compatibility}
\title{check if version indication is compliant.}
\usage{
lib.check_compatibility(condition, version)
}
\arguments{
\item{condition}{A version indication like `>= 4.5.1` or `2.3.4` or `> 1.2.3` or `''` (empty) or `NA`.}

\item{version}{A version number like `1.2.3`, or a vector of version strings (will be converted to `numeric_version('1.2.3') during comparison).}
}
\value{
A logical indicating if the version is considered compatible.
}
\description{
Returns TRUE if the 'condition' complies with the provided 'version'.
This function is vectorized.
}
