% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{lib.dependencies_online}
\alias{lib.dependencies_online}
\title{Check a package his online dependencies}
\usage{
lib.dependencies_online(packageName, cran_url = "https://cran.rstudio.com/")
}
\arguments{
\item{packageName}{The package name to check.}

\item{cran_url}{Defaults to 'https://cran.rstudio.com/'.}
}
\value{
Returns a named character with the packages and their version conditions
  which the given package depends on.
}
\description{
Returns a \code{c(name = '<version spec>')} array which can be used for \code{lib.load()},
\code{lib.install_if_not_compatible()} or \code{lib.dependsOnMe()}.
}
