% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.devTools_install}
\alias{lib.devTools_install}
\title{Installs the incredible `devtools` package.}
\usage{
lib.devTools_install(lib_location = lib.location(), force_install = FALSE)
}
\arguments{
\item{lib_location}{The library (can be an empty folder) to install Devtools in. Defaults to \code{lib.location()}}

\item{force_install}{FALSE by default, if the package `devTools` is already installed, it will return silently.}
}
\description{
Will install devtools and it's dependencies into the multiversion library provided by `lib.location()`.
An alternative library location can optionally be specified.
}
