% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.devtools_load}
\alias{lib.devtools_load}
\title{Loads `devtools` version 1.13.1 and it's dependencies.}
\usage{
lib.devtools_load(lib_location = lib.location())
}
\arguments{
\item{lib_location}{The (version controlled) library to load devtools from.
Use \code{lib.install('devtools', allow_overwrite_on_convert = TRUE)} to install devtools, if you have not done so already.}
}
\value{
No return value, called for it's side-effect of loading the devtools and
  testthat packages. Also the library paths of both packages will be added to the \code{.libPaths()}
}
\description{
During the library call, \code{appendLibPaths} is TRUE, making sure that some devtools functionality
(like running tests) in child R instances will still work and know where to load their libraries from.
}
