% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.location}
\alias{lib.location}
\title{The R_MV_library location.}
\usage{
lib.location(set_session_path)
}
\arguments{
\item{set_session_path}{(optional) If no environment variable has been set to indicate the library location,
You can call this function and let it set the environment variable for this session only.}
}
\value{
When a path is provided, this will be stored as
  the multiversion library location to use during this session (via the environment
  variable "R_MV_LIBRARY_LOCATION"). In all cases, it will return the location of
  the multiversion library. When it cannot be found, it will return an error indicating what to do.
}
\description{
This function will look for the environment variable \code{R_MV_LIBRARY_LOCATION} indicating the R_MV_library location.
Alternatively you can provide a path for this session only, using \code{\link{lib.location}(yourPath)}.
This will set the environment variable for this session.
(You might want to consider to add this to your \code{.Rprofile} file, see \code{?Startup})
}
