% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{lib.package_version_loaded}
\alias{lib.package_version_loaded}
\title{Check the versions of an already loaded package.}
\usage{
lib.package_version_loaded(packageNames, exclude_not_loaded = TRUE)
}
\arguments{
\item{packageNames}{The name or a vector of names of the packages for which to obtain the version.}

\item{exclude_not_loaded}{If true, the default, it will not try to find a 'loaded version' of a package that is not loaded.}
}
\value{
A named character vector with the package name(s) and it's version that is loaded.
}
\description{
This works for both packages within the R_MV_library (which are loaded) and
for packages outside the library.
}
