% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.packs_str2vec}
\alias{lib.packs_str2vec}
\title{Parse single string to named character vector.}
\usage{
lib.packs_str2vec(deps)
}
\arguments{
\item{deps}{A string of length one with the format shown in the description. This will be converted to a named character vector.}
}
\value{
Returns a character vector with the packages and their versions that it could derive from the single provided string.
}
\description{
Parses a string shaped like: \cr
\code{   assertthat (>= 0.1), R6 (>= 2.1.2), Rcpp (>= 0.12.3), tibble (>= 1.2), magrittr (>= 1.5), lazyeval (>= 0.2.0), DBI (>= 0.4.1)} \cr
to match the normal package name/version layout like: \cr
\code{   assertthat          R6         Rcpp      tibble    magrittr     lazyeval          DBI} \cr
\code{     ">= 0.1"  ">= 2.1.2"  ">= 0.12.3"    ">= 1.2"    ">= 1.5"   ">= 0.2.0"   ">= 0.4.1"} \cr \cr
Used by \code{\link{lib.dependencies_online}} to interpret the by CRAN provided dependencies,
used to parse the 'vc_override_dependencies.txt' files and the dependencies mentioned in the 'DESCRIPTION' files of the installed packages.
}
