% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{lib.packs_vec2str}
\alias{lib.packs_vec2str}
\title{Convert package name/version vector to single string.}
\usage{
lib.packs_vec2str(x, do_return = FALSE)
}
\arguments{
\item{x}{A named character vector with package names/versions.
\code{c(dplyr = '>= 1.5.0', data.table = '')}}

\item{do_return}{If FALSE (the default) the package sting is printed, if TRUE,
it is returned as a character string and not printed.}
}
\value{
When do_return = TRUE, returns a character string that describes a
  vector with packages and their version specifications like \code{"dplyr (>= 1.5.0), data.table"}.
  When FALSE, it prints this string and returns nothing.
}
\description{
Used to print a set of package names and their version criteria in a way that
 \code{lib.packs_str2vec()} can parse it again to a package vector.
This way we can list the dependencies of a function easily and support command line interaction for example.
}
