% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{list_untracked_VC_libs}
\alias{list_untracked_VC_libs}
\title{List all untracked library folders}
\usage{
list_untracked_VC_libs(lib.location = R_VC_library_location())
}
\arguments{
\item{lib.location}{By default the default library path obtained with \code{R_VC_library_location()}.}
}
\description{
List all untracked direcories (libraries) within the R_VC_library. The returned untracked directories are cleaned up
and printed so that only the unique combinations of each library and it's version is shown once.
}

