% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{loadDevtools}
\alias{loadDevtools}
\title{Loads `devtools` version 1.13.1 and it's dependencies.}
\usage{
loadDevtools()
}
\description{
During the library call, `appendLibPaths` is TRUE, making sure that some devtools functionallity
(like running tests) in child R instances will still work and know where to load their libraries from.
}
