% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{loadedPackageVersion}
\alias{loadedPackageVersion}
\title{Check the versions of an already loaded package.}
\usage{
loadedPackageVersion(packageName)
}
\arguments{
\item{packageName}{The name or a vector of names of the packages for which to obtain the version.}
}
\description{
Check the versions of an already loaded package.
}
