% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{printExampleLibCall}
\alias{printExampleLibCall}
\title{Print example \code{library_VC} call.}
\usage{
printExampleLibCall(packNameVersion)
}
\arguments{
\item{packNameVersion}{A named character vector with package names and their version indication (e.g. `c(dplyr = '>= 0.4.0', ggplot = '')`).}
}
\description{
Prints the library call that you can use based on a name/version input vector.
}

