% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{printPackageList}
\alias{printPackageList}
\title{Convert package name/version vector to single string.}
\usage{
printPackageList(x, do_return = FALSE)
}
\arguments{
\item{x}{A named character vector with package names/versions.}
}
\description{
Used to print a set of package names and their version criteria in a way that \code{cleanupDependencyList()} can parse it again to a package list.
This way we can list the dependencies of a function easily and it makes it possible to use it when performing a commandline call.
}
