% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{R_VC_library_location}
\alias{R_VC_library_location}
\title{The R_VC_library location.}
\usage{
R_VC_library_location(set_session_path = NULL)
}
\arguments{
\item{set_session_path}{(optional) If no environment variable has been set to indicate the library location,
You can call this function and let it set the environment variable for this session only.}
}
\description{
This function will look for the environment variable \code{R_VC_LIBRARY_LOCATION} indicating the R_VC_library location.
Alternatively you can provide a path for this session only using \code{R_VC_library_location(yourPath)}.
}

