% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{R_VC_temp_lib_location}
\alias{R_VC_temp_lib_location}
\title{Temporary directory location.}
\usage{
R_VC_temp_lib_location(lib.location = R_VC_library_location())
}
\arguments{
\item{lib.location}{By default the default library path obtained with \code{R_VC_library_location()}.}
}
\description{
Indicates the default directory for initially installing a package before it is 'converted' to the final VC library structure (see: \code{convert_to_VC_library()}).
This folder can be cleaned up using \code{cleanTempInstallFolder()} after installing the package succeeded.
This is not done automatically but won't influence the installation of other packages.
}
