% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_undesired_libPaths}
\alias{remove_undesired_libPaths}
\title{Exclude not relevant search paths.}
\usage{
remove_undesired_libPaths(lib.location = R_VC_library_location(),
  dry.run = FALSE)
}
\arguments{
\item{lib.location}{The folder which contains the R_VC_library structure. All directories in .libPaths containing this path will be kept.
By default, it checks the environment variable \code{R_VC_LIBRARY_LOCATION} to find this directory.}

\item{dry.run}{If TRUE, will return the paths that would remain after cleaning up the .libPaths() list.}
}
\description{
Excludes all .libPaths other then those needed for library_VC().
}
