% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{reset.libPaths}
\alias{reset.libPaths}
\title{Minor wrapper to reset libraries to the provided state.}
\usage{
reset.libPaths(currentLibs)
}
\description{
Minor wrapper to reset libraries to the provided state.
}

