% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{strRemain}
\alias{strRemain}
\title{Removes pattern A and pattern B from a string.}
\usage{
strRemain(patA, patB, str)
}
\arguments{
\item{patA}{The first regex pattern to remove from the string.}

\item{patB}{The second regex pattern to remove from the remaining of the first removal.}

\item{str}{The string that needs cleaning up.}
}
\description{
Removes pattern A and pattern B from a string.
}
