% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helper_functions.R
\name{test_helpers}
\alias{test_helpers}
\alias{.initialize_test}
\alias{.add_test}
\alias{.exp_error}
\alias{.exp_match}
\alias{.exp_equal}
\alias{.exp_false}
\alias{.exp_true}
\alias{.get_tests}
\alias{.run_test_batch}
\title{Functions to help with testing inside seperate processes.}
\usage{
.initialize_test(desc)

.add_test(type, A, B)

.exp_error(expr, exp_msg)

.exp_match(expr, regex)

.exp_equal(expr, ref)

.exp_false(expr)

.exp_true(expr)

.get_tests()

.run_test_batch(test_outcomes)
}
\description{
The functions can be used in the following way:\cr\cr
\enumerate{
\item{Use \code{.initialize_test(desc = 'My first test')} to setup for a new test and create a clean slate.}
\item{Then use the different helper functions \code{.exp_...} to perform your tests just like you would do with \code{testthat::expect_...} functions.}
\item{To give back the test outcomes, since it is a separate process, you must use \code{test_outcomes <- .get_tests()}.}
\item{Finally, when all the tests are done and the results are passed back, perform: \code{.run_test_batch(test_outcomes)}.}
}
}
