% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{bareVersion}
\alias{bareVersion}
\title{Remove `>` or `>=` from version string.}
\usage{
bareVersion(packageVersion)
}
\arguments{
\item{packageVersion}{A version indication you would like to remove `>` and `>=` from.}
}
\description{
Remove `>` or `>=` from version string.
}
