% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{chooseVersion}
\alias{chooseVersion}
\title{Choose version based on the version indication, and available versions.}
\usage{
chooseVersion(packVersion, versionList, packageName = "", pick.last = FALSE)
}
\arguments{
\item{packVersion}{A single named version with package name and it's version like \code{c(dplyr = '>= 0.4.0')}.}

\item{versionList}{A list of available versions for this package to choose from. It is the list to choose from and check availability. Created with \code{availablePackageVersions}.}

\item{packageName}{It is used for clear error handling. It should be the package name it is trying to load so we can mention it when crashing.}

\item{pick.last}{If a version like \code{>= 0.5} is given and multiple versions exist, a choice needs to be made.
By default it will take the same or first higher version (when it exists, just \code{0.5}, which is often the case).
This because this is most likely to not change the behaviour of the code. Alternatively, picking the latest version is most
likely to be accepted by other packages their dependencies (e.g. if a package that is loaded later on depends on this package but asks for \code{> 0.6}, it will crash).
The downside of this is that an update could be a major one, going from \code{0.5} to \code{2.0}, where allot of things can change and code is likely to not work anymore.}
}
\description{
Obtains the correct version based on the version instruction provided (e.g. \code{>= 0.5}), the package name and it's available versions.
If no compatible version is found between the available versions a suitable error is thrown.
All different version indications should be handled in this function, including:
1. a version with \code{>} or \code{>=} indicator
2. a version without \code{>}, so just a version e.g. \code{'0.5.0'} (most specific)
3. a zero length char e.g. \code{''} (auto determine)
}
