% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{dependencies}
\alias{dependencies}
\title{List the dependencies of a package.}
\usage{
dependencies(packageName, lib.location = R_VC_library_location())
}
\arguments{
\item{packageName}{The (unquoted) package name for which you would like to print the dependencies.}
}
\description{
Provide a package name (can be without quotes) to show its dependencies.
To list all dependencies of the complete library, use the inversed function "\code{dependsOnMe(all)}" with the value 'all'.
That function also does not require quotes.
}
