% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dependencies}
\alias{dependencies}
\title{Alias of \code{\link{lib.dependencies}}}
\usage{
dependencies(packageName, do_print = TRUE, character.only = FALSE,
  lib_location = lib.location())
}
\description{
Alias of \code{\link{lib.dependencies}}
}
