% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{detachAll}
\alias{detachAll}
\title{Detach all loaded packages and namespaces.}
\usage{
detachAll(
  reload_multiversion = FALSE,
  dryRun = FALSE,
  packageList = names(sessionInfo()$otherPkgs)
)
}
\arguments{
\item{dryRun}{If TRUE, lists all packages that will be cleaned up.}

\item{packageList}{A character vector with the packages to detach/unload. Defaults to all packages.}
}
\description{
Tries to detach all loaded packages and namespaces. Not always stable (within Rstudio).
A restart of Rstudio might be required since it will often hold on to certain namespaces.
A proper reset of all libraries is not possible, this is the best we can do. \cr
\cr
In general, it is possible to create a complete clean environment by clearing your workspace,
running \code{detachAll} and then restarting Rstudio. If problems with package loading still persists,
then follow the final alternative solution described in the details section of the documentation of \code{lib.load}.
}
