% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_with_packages.R
\name{execute_with_packages}
\alias{execute_with_packages}
\title{Perform operation with a certain set of packages.}
\usage{
execute_with_packages(packages_to_load = NULL, func_handle, ...,
  lib.location = R_VC_library_location(), execute_quietly = FALSE)
}
\arguments{
\item{packages_to_load}{A sting indicating which packages must be loaded (e.g. `"dplyr (0.5.0), ggplot2, tidyr (1.2.3)"`).
It is shaped the way like `depends` field is showing the dependencies of the different packages in `available.packages()`, which is equal to the `DESCRIPTION` file.
It will be parsed using `parse_dependency_string` to become an offten used format looking like: `"c(dplyr = '0.5.0', ggplot2 = '', tidyr = '1.2.3')"`, converting back can be done using `packageList2str()`.}

\item{func_handle}{A function object or the function name as a character string.}

\item{...}{Provide all the remaining arguments which will be arguments for the function handle.}

\item{lib.location}{The location of the version controlled library. Defaults to R_VC_library_location(), which is the directory provided by the environment variable.}

\item{execute_quietly}{The complete process content will be printed when the process is finished. If FALSE, only the returned file name will be printed.}
}
\description{
This function can be used to perform some R operation with a configured set of loaded packages.
}
