% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{install.packages_VC}
\alias{install.packages_VC}
\title{Install CRAN package into VC library.}
\usage{
install.packages_VC(installPackages = NULL,
  lib.location = R_VC_library_location(), add_to_VC_library = TRUE)
}
\arguments{
\item{installPackages}{Provide a vector of package names. A version cannot be supplied.}

\item{lib.location}{The folder containing a structure where this package can be installed.
The package will first be test installed in a temporary install folder \code{TEMP_install_location}
indicated by the \code{R_VC_temp_lib_location} function. If \code{add_to_VC_library} is set to TRUE (the default).
the installed package(s) is moved to the destination location automatically.}

\item{add_to_VC_library}{If TRUE (the default), the installed package(s) is (are) moved to the final destination automatically.
Otherwise it is necessary to run \code{convert_to_VC_library()} manually after the installation into the temporary folder finished.}
}
\description{
This function can be used to install a CRAN package in the (default or specified) R_VC_library.
For installing packages from source use \code{install.packages_VC_tarball}.
}
