% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{install.packages_VC}
\alias{install.packages_VC}
\title{Alias for \code{\link{lib.install}}}
\usage{
install.packages_VC(installPackages = NULL, lib_location = lib.location(),
  install_temporarily = FALSE, overwrite_this_package = FALSE,
  execute_with_Rscript = TRUE, verbose = FALSE,
  cran_url = "http://cran.us.r-project.org")
}
\description{
Alias for \code{\link{lib.install}}
}

