% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{installed.packages_VC}
\alias{installed.packages_VC}
\title{Alias of \code{\link{lib.installed_packages}}}
\usage{
installed.packages_VC(lib_location = lib.location())
}
\description{
Alias of \code{\link{lib.installed_packages}}
}
