% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.clean_libPaths}
\alias{lib.clean_libPaths}
\title{Exclude not relevant search paths.}
\usage{
lib.clean_libPaths(lib_location = lib.location(), dry.run = FALSE)
}
\arguments{
\item{lib_location}{The folder which contains the multiversion library. All directories in .libPaths containing this path will be kept.
By default, it checks the environment variable \code{R_MV_LIBRARY_LOCATION} to find this directory.}

\item{dry.run}{If TRUE, will not change the paths but will print the paths that would remain after cleaning up the .libPaths() list.}
}
\description{
Excludes all .libPaths other then those needed for lib.load().
}
