% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{lib.dependencies_online}
\alias{lib.dependencies_online}
\title{Check if an package belongs to the standard R packages}
\usage{
lib.dependencies_online(packageName, cran_url = "https://cran.rstudio.com/")
}
\arguments{
\item{packageName}{The package name to check.}

\item{cran_url}{Defaults to 'https://cran.rstudio.com/'.}
}
\description{
Check if an package belongs to the standard R packages
}

