% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.devtools_load}
\alias{lib.devtools_load}
\title{Loads `devtools` version 1.13.1 and it's dependencies.}
\usage{
lib.devtools_load(lib_location = lib.location())
}
\arguments{
\item{lib_location}{The (version controlled) library to load devtools from.
Use lib.devTools_install to install devtools, if you have not done so already.}
}
\description{
During the library call, `appendLibPaths` is TRUE, making sure that some devtools functionallity
(like running tests) in child R instances will still work and know where to load their libraries from.
}

