% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_with_packages.R
\name{lib.execute_using_packagelist}
\alias{lib.execute_using_packagelist}
\title{Perform operation with a certain set of packages.}
\usage{
lib.execute_using_packagelist(packages_to_load = NULL, func_handle, ...,
  execution_log_location = normPath(paste0("./execution_logs/execution_",
  gsub("\\\\s|-|:", "_", format(Sys.time())), ".txt")),
  lib_location = lib.location(), wait_for_response = TRUE,
  execute_quietly = FALSE)
}
\arguments{
\item{packages_to_load}{A sting indicating which packages must be loaded (e.g. `"dplyr (0.5.0), ggplot2, tidyr (1.2.3)"`).
It is shaped the way like `depends` field is showing the dependencies of the different packages in `available.packages()`, which is equal to the `DESCRIPTION` file.
It will be parsed using `lib.packs_str2vec` to become an offten used format looking like: `"c(dplyr = '0.5.0', ggplot2 = '', tidyr = '1.2.3')"`, converting back can be done using `lib.packs_vec2str()`.}

\item{func_handle}{A function object or the function name as a character string.}

\item{...}{Provide all the remaining arguments which will be arguments for the function handle.}

\item{lib_location}{The location of the version controlled library. Defaults to lib.location(), which is the directory provided by the environment variable.}

\item{execute_quietly}{The complete process content will be printed when the process is finished. If FALSE, only the returned file name will be printed.}
}
\description{
This function can be used to perform some R operation with a configured set of loaded packages.
}

