% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.git_show_untracked}
\alias{lib.git_show_untracked}
\title{List all untracked library folders}
\usage{
lib.git_show_untracked(lib_location = lib.location())
}
\arguments{
\item{lib_location}{By default the default library path obtained with \code{lib.location()}.}
}
\description{
List all untracked direcories (libraries) within the multiversion library. The returned untracked directories are cleaned up
and printed so that only the unique combinations of each library and it's version is shown once.
}

