% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{lib.installed_packages}
\alias{lib.installed_packages}
\title{Show the complete library content.}
\usage{
lib.installed_packages(lib_location = lib.location())
}
\description{
Use to print all available packages in the VC_library with all their versions including their dependencies.
Simply performs a call to \code{lib.dependsOnMe(all)}.
}
