% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lib.location_install_dir}
\alias{lib.location_install_dir}
\title{Temporary directory location.}
\usage{
lib.location_install_dir(lib_location = lib.location())
}
\arguments{
\item{lib_location}{By default the default library path obtained with \code{lib.location()}.}
}
\description{
Indicates the default directory for initially installing a package before it is 'converted' to the final VC library structure (see: \code{lib.convert()}).
This folder can be cleaned up using \code{cleanTempInstallFolder()} after installing the package succeeded.
This is not done automatically but won't influence the installation of other packages.
}
