% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version_management.R
\name{lib.package_version_loaded}
\alias{lib.package_version_loaded}
\title{Check the versions of an already loaded package.}
\usage{
lib.package_version_loaded(packageNames, dont_exclude_not_loaded = F)
}
\arguments{
\item{packageName}{The name or a vector of names of the packages for which to obtain the version.}
}
\description{
Check the versions of an already loaded package.
}

