% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{normPath}
\alias{normPath}
\title{Normalize path with backslashes.}
\usage{
normPath(path)
}
\arguments{
\item{path}{The path which needs to be normalized. Will make `C:/PROGRA~1/R/R-33~1.1/library` into `C:/Program Files/R/R-3.3.1/library`.}
}
\description{
This short-hand function normalizes the path and makes sure only forward slashes are used.
Other slashes are not usable in `grepl` statements directly for example, the '\\' is parsed to '\' before being used as regex.
}

