% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VC_library_writing.R
\name{tryLoadPackages_VC}
\alias{tryLoadPackages_VC}
\title{Test load package list.}
\usage{
tryLoadPackages_VC(packages, lib.location)
}
\description{
Internally used for 'test loading' a list of dependencies when installing a package.
Returns a list of succesfull load operations.
}
